refactor: separate agent transport and player state actions
Extract Unix-domain socket connection, watchdog, and wire handling from typed RPC mapping. Split synthetic-player state, message, inventory, and drop actions from lifecycle and world interactions, with direct dispatcher dependencies and no forwarding facades. Addresses M18.
