+++ lightkeeper-agent-spigot/src/main/java/nl/pim16aap2/lightkeeper/agent/spigot/AgentPlayerActions.java
@@ -2,12 +2,7 @@ package nl.pim16aap2.lightkeeper.agent.spigot;
 
 import nl.pim16aap2.lightkeeper.nms.api.IBotPlayerNmsAdapter;
 import nl.pim16aap2.lightkeeper.protocol.CreatePlayer;
-import nl.pim16aap2.lightkeeper.protocol.DropItem;
 import nl.pim16aap2.lightkeeper.protocol.ExecutePlayerCommand;
-import nl.pim16aap2.lightkeeper.protocol.GetPlayerChatComponents;
-import nl.pim16aap2.lightkeeper.protocol.GetPlayerInventory;
-import nl.pim16aap2.lightkeeper.protocol.ItemSnapshot;
-import nl.pim16aap2.lightkeeper.protocol.GetPlayerMessages;
 import nl.pim16aap2.lightkeeper.protocol.LeftClickBlock;
 import nl.pim16aap2.lightkeeper.protocol.PlacePlayerBlock;
 import nl.pim16aap2.lightkeeper.protocol.RemovePlayer;
@@ -19,26 +14,20 @@ import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
-import org.bukkit.entity.Item;
 import org.bukkit.entity.Player;
 import org.bukkit.event.block.Action;
-import org.bukkit.event.player.PlayerDropItemEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.java.JavaPlugin;
-import tools.jackson.databind.ObjectMapper;
-
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
 
 /**
- * Protocol action handler for synthetic player lifecycle and player-driven interactions.
+ * Protocol action handler for synthetic player lifecycle, movement, and world interactions.
  *
  * <p>This class is responsible for creating/removing synthetic players, executing commands as those players,
- * placing blocks through player context, and returning captured player-facing messages.
+ * placing blocks through player context, and teleporting players.
  */
 final class AgentPlayerActions
 {
@@ -55,10 +44,6 @@ final class AgentPlayerActions
      */
     private final AgentSyntheticPlayerStore playerStore;
     /**
-     * JSON mapper used to serialize inventory snapshot payloads.
-     */
-    private final ObjectMapper objectMapper;
-    /**
      * NMS-backed synthetic player implementation.
      */
     private final IBotPlayerNmsAdapter botPlayerNmsAdapter;
@@ -70,8 +55,6 @@ final class AgentPlayerActions
      *     Main-thread execution bridge for Bukkit-safe operations.
      * @param playerStore
      *     Registry containing synthetic players and related state.
-     * @param objectMapper
-     *     JSON serializer for inventory snapshots.
      * @param botPlayerNmsAdapter
      *     NMS adapter used to spawn/remove synthetic players and drain received messages.
      */
@@ -79,13 +62,11 @@ final class AgentPlayerActions
         JavaPlugin plugin,
         AgentMainThreadExecutor mainThreadExecutor,
         AgentSyntheticPlayerStore playerStore,
-        ObjectMapper objectMapper,
         IBotPlayerNmsAdapter botPlayerNmsAdapter)
     {
         this.plugin = Objects.requireNonNull(plugin, "plugin");
         this.mainThreadExecutor = Objects.requireNonNull(mainThreadExecutor, "mainThreadExecutor");
         this.playerStore = Objects.requireNonNull(playerStore, "playerStore");
-        this.objectMapper = Objects.requireNonNull(objectMapper, "objectMapper");
         this.botPlayerNmsAdapter = Objects.requireNonNull(botPlayerNmsAdapter, "botPlayerNmsAdapter");
     }
 
@@ -270,143 +251,6 @@ final class AgentPlayerActions
             command.uuid(), command.x(), command.y(), command.z(), command.blockFace(), Action.RIGHT_CLICK_BLOCK
         );
         return new RightClickBlock.Response(cancelled);
-    }
-
-    /**
-     * Handles {@code GET_PLAYER_MESSAGES} by draining adapter messages and returning full tracked history.
-     *
-     * @param command
-     *     Typed command carrying the player UUID.
-     * @return Response with the accumulated message list.
-     *
-     * @throws Exception
-     *     Propagates main-thread execution failures.
-     */
-    GetPlayerMessages.Response handleGetPlayerMessages(GetPlayerMessages.Command command)
-        throws Exception
-    {
-        final UUID uuid = command.uuid();
-        final List<String> messages = mainThreadExecutor.callOnMainThread(() ->
-        {
-            playerStore.getRequiredPlayer(uuid);
-            playerStore.capturePlayerMessages(botPlayerNmsAdapter, uuid);
-            return playerStore.getPlayerMessages(uuid);
-        });
-
-        return new GetPlayerMessages.Response(messages);
-    }
-
-    /**
-     * Handles {@code GET_PLAYER_CHAT_COMPONENTS} by returning accumulated component JSON payloads.
-     *
-     * @param command
-     *     Typed command carrying the player UUID.
-     * @return Response containing the serialized chat components.
-     * @throws Exception
-     *     Propagates main-thread execution and serialization failures.
-     */
-    GetPlayerChatComponents.Response handleGetPlayerChatComponents(GetPlayerChatComponents.Command command)
-        throws Exception
-    {
-        final UUID uuid = command.uuid();
-        final String componentsJson = mainThreadExecutor.callOnMainThread(() ->
-        {
-            playerStore.getRequiredPlayer(uuid);
-            playerStore.capturePlayerChatComponents(botPlayerNmsAdapter, uuid);
-            return objectMapper.writeValueAsString(playerStore.getPlayerChatComponents(uuid));
-        });
-
-        return new GetPlayerChatComponents.Response(componentsJson);
-    }
-
-    /**
-     * Handles {@code GET_PLAYER_INVENTORY} by returning a snapshot of the player's inventory.
-     *
-     * @param command
-     *     Typed command carrying the player UUID.
-     * @return Response with {@code inventoryJson}.
-     *
-     * @throws Exception
-     *     Propagates main-thread execution failures.
-     */
-    GetPlayerInventory.Response handleGetPlayerInventory(GetPlayerInventory.Command command)
-        throws Exception
-    {
-        final UUID uuid = command.uuid();
-        final List<ItemSnapshot> items = mainThreadExecutor.callOnMainThread(() ->
-        {
-            final Player player = playerStore.getRequiredPlayer(uuid);
-            return buildInventoryItems(player.getInventory().getContents());
-        });
-
-        return new GetPlayerInventory.Response(items);
-    }
-
-    private static List<ItemSnapshot> buildInventoryItems(ItemStack... contents)
-    {
-        final List<ItemSnapshot> items = new ArrayList<>();
-        for (int i = 0; i < contents.length; i++)
-        {
-            final ItemStack item = contents[i];
-            if (item == null || AgentMaterials.isAir(item.getType()))
-                continue;
-            items.add(AgentItemSnapshots.of(i, item));
-        }
-        return items;
-    }
-
-    /**
-     * Handles {@code DROP_ITEM} by dropping the player's main-hand item into the world.
-     *
-     * <p>Fires a {@code PlayerDropItemEvent}. If cancelled, the item entity is removed and inventory is unchanged.
-     * If not cancelled, the entity stays in the world and one item is consumed from the player's main hand.
-     *
-     * @param command
-     *     Typed command carrying the player UUID.
-     * @return Response indicating whether the drop materialised.
-     *
-     * @throws Exception
-     *     Propagates main-thread execution failures.
-     */
-    DropItem.Response handleDropItem(DropItem.Command command)
-        throws Exception
-    {
-        final UUID uuid = command.uuid();
-        final Boolean dropped = mainThreadExecutor.callOnMainThread(() ->
-        {
-            final Player player = playerStore.getRequiredPlayer(uuid);
-            final ItemStack item = player.getInventory().getItemInMainHand();
-            if (item == null || AgentMaterials.isAir(item.getType()))
-                return Boolean.FALSE;
-
-            final ItemStack singleItem = item.clone();
-            singleItem.setAmount(1);
-            final Item droppedItem = player.getWorld().dropItemNaturally(player.getLocation(), singleItem);
-            final PlayerDropItemEvent event = new PlayerDropItemEvent(player, droppedItem);
-
-            Bukkit.getPluginManager().callEvent(event);
-
-            if (event.isCancelled())
-            {
-                droppedItem.remove();
-                return Boolean.FALSE;
-            }
-
-            // Consume one item from the player's main hand
-            if (item.getAmount() > 1)
-            {
-                item.setAmount(item.getAmount() - 1);
-                player.getInventory().setItemInMainHand(item);
-            }
-            else
-            {
-                player.getInventory().setItemInMainHand(null);
-            }
-
-            return Boolean.TRUE;
-        });
-
-        return new DropItem.Response(dropped);
     }
 
     /**
